Addition: Made the pen annotation roll over (before it has been saved in hte database)
Addition: Made the IDGUI button/toggle button include options to set their visibility, and inclusion in layout

Bug Fix: Fixed up the browser toolbar so the upload button only shows in fixed collections, and in regular collections where the user has the modify permission.

Bug Fix: The annotation list was not matching the user (to establish the delete permissions).

Bug Fix: Deleting a object that had only 1 user with access to it.

Signed-off-by: Craig McNamara <cmcnamara87@gmail.com>