Added Close Button to Shelf
Added Refresh timer to collection list (60 seconds)

Removed Share Panel from Media View

Added Portals panel to Media View - shows portals the asset is in.

FIxed uploading an asset to the collection + sharing

Added Loading Icon (not used as of yet)

Added Database call to get names of collections

Added XMLList of users that have access to media

Added Copying a Collections Access to assets in the collection

Incorporated collection copying into other areas.

Renamed sharing panel in Browser to be 'Add People'

Made the delete button for a comment, only appear if the comment hasnt been changed to 'comment removed' or if the logged in user is the sys-admin

Made the new comments disable themselves while saving

Created the Portal panel (displayed in media view)

Signed-off-by: Craig McNamara <cmcnamara87@gmail.com>